Refactor components into modules
